cleaner code and added cave man style
